Fix badges not showing, remove DM unread badges
When you start discord and don't select a channel and that channel is unread `getUnreadCount` will still return `0`.
This caused the indicator to not be shown on channels you did not select since discord started. Now it just shows an empty badge.
The DM unread indicator got removed (from DMs and group DMs) because unread messages and mentions are the same for those.